+++ src/main/java/roomescape/admin/controller/AdminController.java
@@ -1,4 +1,4 @@
-package roomescape.controller;
+package roomescape.admin.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
