+++ src/main/java/roomescape/reservation/repository/ReservationRepository.java
@@ -1,8 +1,8 @@
-package roomescape.repository;
+package roomescape.reservation.repository;
 
 import java.util.List;
 import java.util.Optional;
-import roomescape.domain.Reservation;
+import roomescape.reservation.domain.Reservation;
 
 public interface ReservationRepository {
