+++ src/test/java/roomescape/admin/controller/AdminControllerTest.java
@@ -1,6 +1,4 @@
-package roomescape;
-
-import static org.hamcrest.Matchers.is;
+package roomescape.admin.controller;
 
 import io.restassured.RestAssured;
 import org.junit.jupiter.api.DisplayName;
@@ -10,7 +8,7 @@ import org.springframework.test.annotation.DirtiesContext;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
-public class AdminTest {
+public class AdminControllerTest {
 
     @Test
     @DisplayName("어드민 페이지를 반환한다.")
@@ -22,17 +20,11 @@ public class AdminTest {
     }
 
     @Test
-    @DisplayName("모든 예약을 조회한다.")
+    @DisplayName("모든 예약 페이지를 반환한다.")
     void getAllReservations() {
         RestAssured.given().log().all()
                 .when().get("/admin/reservation")
                 .then().log().all()
                 .statusCode(200);
-
-        RestAssured.given().log().all()
-                .when().get("/reservations")
-                .then().log().all()
-                .statusCode(200)
-                .body("size()", is(0));
     }
 }
